+++ src/main/java/moonfather/lilypads/mixin/falling/DamageCancelWater.java
@@ -12,7 +12,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class DamageCancelWater extends DamageCancelBase
 {
     @Override
-    public void cancelDamage(Level world, BlockState state, BlockPos pos, Entity entity, float fallDistance, CallbackInfo ci)
+    public void cancelDamage(Level world, BlockState state, BlockPos pos, Entity entity, double fallDistance, CallbackInfo ci)
     {
         entity.fallDistance = 0.1f;
         ci.cancel();
